refactor: move `animation.py` from `anims` to `anims_core`
