+++ janim/anims/transform.py
@@ -11,7 +11,7 @@ from typing import Callable, Generator, Iterable
 
 import numpy as np
 
-from janim.anims.animation import Animation
+from janim.anims_core.animation import Animation
 from janim.anims.composition import AnimGroup
 from janim.anims.fading import FadeIn, FadeInFromPoint, FadeOut, FadeOutToPoint
 from janim.anims_core.stackable import ApplyParams, ItemAnimation
