+++ janim/anims/updater.py
@@ -9,9 +9,9 @@ from typing import Any, Callable, Iterable, Self
 
 from tqdm import tqdm as ProgressDisplay
 
-from janim.anims.animation import FOREVER, Animation
 from janim.anims.method_updater_meta import METHOD_UPDATER_KEY, MethodUpdaterInfo
 from janim.anims_core.anim_stack import AnimStack
+from janim.anims_core.animation import FOREVER, Animation
 from janim.anims_core.stackable import ApplyAligner, ApplyParams, ItemAnimation
 from janim.anims_core.time import TimeRange
 from janim.components.component import Component
