+++ janim/gui/timeline_view.py
@@ -9,7 +9,7 @@ from PySide6.QtCore import QPoint, QPointF, QRect, Qt, QTimer, Signal
 from PySide6.QtGui import QColor, QKeyEvent, QMouseEvent, QPainter, QPaintEvent, QPen, QWheelEvent
 from PySide6.QtWidgets import QLabel, QVBoxLayout, QWidget
 
-from janim.anims.animation import FOREVER, Animation
+from janim.anims_core.animation import FOREVER, Animation
 from janim.anims.composition import AnimGroup
 from janim.anims.timeline import BuiltTimeline, Timeline, TimelineItem
 from janim.anims_core.time import TimeRange
